+++ app/src/main/java/woowacourse/payments/ui/add/AddPaymentCardActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui
+package woowacourse.payments.ui.add
 
 import android.content.Context
 import android.content.Intent
