+++ app/src/main/java/woowacourse/payments/ui/cards/PaymentCardsScreen.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui
+package woowacourse.payments.ui.cards
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
